move StartingFromSnapshotStage to the existing internal package
