+++ core/src/test/scala/akka/persistence/r2dbc/query/StartingFromSnapshotStageSpec.scala
@@ -13,7 +13,7 @@ import akka.actor.typed.ActorSystem
 import akka.persistence.Persistence
 import akka.persistence.query.TimestampOffset
 import akka.persistence.query.typed.EventEnvelope
-import akka.persistence.r2dbc.query.internal.StartingFromSnapshotStage
+import akka.persistence.r2dbc.internal.StartingFromSnapshotStage
 import akka.persistence.typed.PersistenceId
 import akka.stream.scaladsl.Source
 import akka.stream.testkit.scaladsl.TestSink
